+++ src/type/ApexClassPermission.ts
@@ -1,5 +1,0 @@
-export type ApexClassPermission = {
-    id: string;
-    permissionId: string;
-    apexClassId: string;
-};