+++ src/type/CustomSettingPermission.ts
@@ -1,5 +1,0 @@
-export type CustomSettingPermission = {
-    id: string;
-    permissionId: string;
-    customSettingId: string;
-};