+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.6.0
+
+- Set permissions for custom metadata types
+- Set permissions for custom permissions
+- Set permissions for visualforce pages
+- Internal structure changes
+- Bug fixes
+
 ## 0.5.0
 
 - Set permissions for custom setting